Este commit implementa a funcionalidade principal da API de e-commerce, incluindo correções de bugs críticas, refatoração de arquitetura e implementação de funcionalidades de negócio essenciais.
Principais Alterações:

Refatoração de Pedidos (CRUD):

Implementado POST /api/pedido (criar), GET /api/pedido (listar), GET /api/pedido/{id} (buscar) e PUT /api/pedido/{id}/status/{idStatus} (atualizar status).

Corrigida a lógica inicial no PedidoService, que usava getters em vez de setters para salvar dados.

Refatorado PedidoController para usar PedidoService em vez de injeção direta do repositório.

Validação e Gestão de Erros:

Adicionada a dependência spring-boot-starter-validation.

Implementada validação (@Valid, @NotNull, @Min) nos DTOs de entrada (PedidoRequestDTO, ItemPedidoRequestDTO).

Criado um GlobalExceptionHandler (@RestControllerAdvice) para formatar exceções de validação e de RuntimeException em respostas JSON padronizadas.

Funcionalidades de "Smart Delete":

ClienteService e ProdutoService agora verificam se existem pedidos associados antes de permitir a exclusão, evitando erros 500 da base de dados (ON DELETE RESTRICT).

Os endpoints DELETE agora retornam HTTP 400 Bad Request com uma mensagem de erro clara.

Novos Endpoints de Consulta:

Implementado GET /api/pedido/cliente/{clienteId} para listar pedidos por cliente.

Implementado GET /api/produto/categoria/{categoria} para filtrar produtos por categoria.

Performance (Paginação):

Implementado Pageable no endpoint GET /api/produto para suportar paginação, melhorando drasticamente a performance em grandes volumes de dados.

Correções de Bugs e Otimização:

Corrigida a incompatibilidade de tipos de ID (Long vs. Integer) entre as entidades Java e o SERIAL do PostgreSQL.

Corrigido o bug de serialização JSON (ByteBuddyInterceptor) ajustando o nome do getter getIdProduto em ItemPedido.java.

Removidos ItemPedidoController e ItemPedidoService por serem desnecessários, tratando ItemPedido como uma entidade filha do agregado Pedido.

+++ src/main/java/br/byteshop/ecommerce/bs_api/model/Pedido.java
@@ -14,7 +14,7 @@ public class Pedido {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_pedido", nullable = false)
-    private Long id_pedido;
+    private Integer id_pedido;
 
     @Column(name = "numero_pedido", nullable = false, unique = true, length = 50)
     private String numeroPedido;
@@ -83,11 +83,11 @@ public class Pedido {
         this.numeroPedido = numeroPedido;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id_pedido;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id_pedido = id;
     }
 }
